# contracts/openapi.json
{
  "components": {
    "schemas": {
      "ActionBody": {
        "properties": {
          "actor": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Actor",
            "type": "string"
          },
          "note": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "required": [
          "actor"
        ],
        "title": "ActionBody",
        "type": "object"
      },
      "AlertMarker": {
        "properties": {
          "icu_hour": {
            "title": "Icu Hour",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "raised_at": {
            "title": "Raised At",
            "type": "string"
          },
          "source": {
            "enum": [
              "model",
              "news2"
            ],
            "title": "Source",
            "type": "string"
          },
          "status": {
            "enum": [
              "open",
              "acknowledged",
              "escalated",
              "resolved"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "icu_hour",
          "raised_at",
          "source",
          "status"
        ],
        "title": "AlertMarker",
        "type": "object"
      },
      "AlertView": {
        "properties": {
          "calibrated_probability": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Calibrated Probability"
          },
          "encounter_id": {
            "title": "Encounter Id",
            "type": "string"
          },
          "etag": {
            "title": "Etag",
            "type": "string"
          },
          "icu_hour": {
            "title": "Icu Hour",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "model_version": {
            "title": "Model Version",
            "type": "string"
          },
          "mrn": {
            "title": "Mrn",
            "type": "string"
          },
          "news2": {
            "$ref": "#/components/schemas/News2"
          },
          "raised_at": {
            "title": "Raised At",
            "type": "string"
          },
          "raw_score": {
            "title": "Raw Score",
            "type": "number"
          },
          "source": {
            "enum": [
              "model",
              "news2"
            ],
            "title": "Source",
            "type": "string"
          },
          "status": {
            "enum": [
              "open",
              "acknowledged",
              "escalated",
              "resolved"
            ],
            "title": "Status",
            "type": "string"
          },
          "top_factors": {
            "items": {
              "$ref": "#/components/schemas/Factor"
            },
            "title": "Top Factors",
            "type": "array"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          },
          "updated_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated By"
          }
        },
        "required": [
          "id",
          "mrn",
          "encounter_id",
          "source",
          "status",
          "raised_at",
          "icu_hour",
          "raw_score",
          "calibrated_probability",
          "news2",
          "top_factors",
          "model_version",
          "updated_at",
          "updated_by",
          "etag"
        ],
        "title": "AlertView",
        "type": "object"
      },
...
      "ErrorBody": {
        "properties": {
          "alert": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AlertView"
              },
              {
                "type": "null"
              }
            ]
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "EscalateBody": {
        "properties": {
          "actor": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Actor",
            "type": "string"
          },
          "note": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "reason": {
            "enum": [
              "clinical_deterioration",
              "senior_review_needed",
              "sepsis_pathway_started",
              "critical_care_outreach",
              "other"
            ],
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "actor",
          "reason"
        ],
        "title": "EscalateBody",
        "type": "object"
      },
      "Factor": {
        "properties": {
          "contribution": {
            "title": "Contribution",
            "type": "number"
          },
          "feature": {
            "title": "Feature",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "unit": {
            "title": "Unit",
            "type": "string"
          },
          "value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          }
        },
        "required": [
          "feature",
          "label",
          "unit",
          "value",
          "contribution"
        ],
        "title": "Factor",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "News2": {
        "properties": {
          "components": {
            "$ref": "#/components/schemas/News2Components"
          },
          "single_parameter_three": {
            "title": "Single Parameter Three",
            "type": "boolean"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "total",
          "components",
          "single_parameter_three"
        ],
        "title": "News2",
        "type": "object"
      },
      "News2Components": {
        "properties": {
          "consciousness": {
            "title": "Consciousness",
            "type": "integer"
          },
          "pulse": {
            "title": "Pulse",
            "type": "integer"
          },
          "respiratory_rate": {
            "title": "Respiratory Rate",
            "type": "integer"
          },
          "spo2": {
            "title": "Spo2",
            "type": "integer"
          },
          "supplemental_o2": {
            "title": "Supplemental O2",
            "type": "integer"
          },
          "systolic_bp": {
            "title": "Systolic Bp",
            "type": "integer"
          },
          "temperature": {
            "title": "Temperature",
            "type": "integer"
          }
        },
        "required": [
          "respiratory_rate",
          "spo2",
          "supplemental_o2",
          "systolic_bp",
          "pulse",
          "consciousness",
          "temperature"
        ],
        "title": "News2Components",
        "type": "object"
      },
      "PatientVitals": {
        "properties": {
          "admitted_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Admitted At"
          },
          "alerts": {
            "items": {
              "$ref": "#/components/schemas/AlertMarker"
            },
            "title": "Alerts",
            "type": "array"
          },
          "bed": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bed"
          },
          "encounter_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Encounter Id"
          },
          "mrn": {
            "title": "Mrn",
            "type": "string"
          },
          "name": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PersonName"
              },
              {
                "type": "null"
              }
            ]
          },
          "patient_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Patient Id"
          },
          "scores": {
            "items": {
              "$ref": "#/components/schemas/ScorePoint"
            },
            "title": "Scores",
            "type": "array"
          },
          "series": {
            "additionalProperties": {
              "$ref": "#/components/schemas/Series"
            },
            "title": "Series",
            "type": "object"
          },
          "window": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VitalsWindow"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "mrn",
          "encounter_id",
          "series",
          "scores",
          "alerts"
        ],
        "title": "PatientVitals",
        "type": "object"
      },
      "PersonName": {
        "properties": {
          "family": {
            "title": "Family",
            "type": "string"
          },
          "given": {
            "title": "Given",
            "type": "string"
          }
        },
        "required": [
          "family",
          "given"
        ],
        "title": "PersonName",
        "type": "object"
      },
      "ScorePoint": {
        "properties": {
          "alerted": {
            "title": "Alerted",
            "type": "boolean"
          },
          "calibrated_probability": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Calibrated Probability"
          },
          "hour_ending": {
            "title": "Hour Ending",
            "type": "string"
          },
          "icu_hour": {
            "title": "Icu Hour",
            "type": "integer"
          },
          "model_version": {
            "title": "Model Version",
            "type": "string"
          },
          "news2": {
            "$ref": "#/components/schemas/News2"
          },
          "news2_total": {
            "title": "News2 Total",
            "type": "integer"
          },
          "raw_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Score"
          }
        },
        "required": [
          "icu_hour",
          "hour_ending",
          "news2_total",
          "news2",
          "raw_score",
          "calibrated_probability",
          "model_version",
          "alerted"
        ],
        "title": "ScorePoint",
        "type": "object"
      },
      "Series": {
        "properties": {
          "display": {
            "title": "Display",
            "type": "string"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/SeriesPoint"
            },
            "title": "Points",
            "type": "array"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit"
          }
        },
        "required": [
          "display",
          "unit",
          "points"
        ],
        "title": "Series",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "VitalsWindow": {
        "properties": {
          "end": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End"
          },
          "hours": {
            "title": "Hours",
            "type": "integer"
          },
          "start": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start"
          }
        },
        "required": [
          "start",
          "end",
          "hours"
        ],
        "title": "VitalsWindow",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "WardWatch FHIR service",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/alerts": {
      "get": {
        "operationId": "list_alerts_api_alerts_get",
        "parameters": [
          {
            "description": "comma-separated statuses",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "comma-separated statuses",
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AlertView"
                  },
                  "title": "Response List Alerts Api Alerts Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Bad Request"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Alerts"
      }
    },
    "/api/alerts/{alert_id}": {
      "get": {
        "operationId": "get_alert_api_alerts__alert_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "alert_id",
            "required": true,
            "schema": {
              "title": "Alert Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertView"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Alert"
      }
    },
    "/api/alerts/{alert_id}/acknowledge": {
      "post": {
        "operationId": "acknowledge_api_alerts__alert_id__acknowledge_post",
        "parameters": [
          {
            "in": "path",
            "name": "alert_id",
            "required": true,
            "schema": {
              "title": "Alert Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "if-match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertView"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Conflict"
          },
          "412": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Precondition Failed"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "428": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Precondition Required"
          }
        },
        "summary": "Acknowledge"
      }
    },
    "/api/alerts/{alert_id}/escalate": {
      "post": {
        "operationId": "escalate_api_alerts__alert_id__escalate_post",
        "parameters": [
          {
            "in": "path",
            "name": "alert_id",
            "required": true,
            "schema": {
              "title": "Alert Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "if-match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EscalateBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertView"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Conflict"
          },
          "412": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Precondition Failed"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "428": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Precondition Required"
          }
        },
        "summary": "Escalate"
      }
    },
    "/api/alerts/{alert_id}/resolve": {
      "post": {
        "operationId": "resolve_api_alerts__alert_id__resolve_post",
        "parameters": [
          {
            "in": "path",
            "name": "alert_id",
            "required": true,
            "schema": {
              "title": "Alert Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "if-match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertView"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Conflict"
          },
          "412": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Precondition Failed"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "428": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Precondition Required"
          }
        },
        "summary": "Resolve"
      }
    },
    "/api/patients/{mrn}/vitals": {
      "get": {
        "operationId": "vitals_api_patients__mrn__vitals_get",
        "parameters": [
          {
            "in": "path",
            "name": "mrn",
            "required": true,
            "schema": {
              "title": "Mrn",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "hours",
            "required": false,
            "schema": {
              "default": 12,
              "maximum": 168,
              "minimum": 1,
              "title": "Hours",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientVitals"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Vitals"
      }
    },
...
  }
}